refactor(git-diff): move the git layer out of index.js
index.js had grown to 655 lines covering three unrelated jobs: running git,
deciding what to render, and wiring the plugin api. The git half now lives in
plugins/git-diff/git.js, leaving 461 lines in index.js and 225 in the new file.

Moved verbatim: runGit and the trusted and untrusted wrappers around it, the git
environment, resolveRepo, hasCommits, resolveBase, listWorktrees, assessTrust and
the numstat reader.

Two seams changed while moving. untrackedPatch became listUntracked, which
returns the sorted paths and nothing else, so turning them into a patch stays
with the code that assembles the diff rather than with the code that talks to
git. tooBigFileList split for the same reason: numstatFiles returns the tracked
side, and untrackedFileList in index.js supplies the rest from records the scan
already holds.

parseNumstat is importable now, so the record shapes it reads are asserted
directly rather than only through a rendered diff: an ordinary record, a rename
with its empty path field and two following records, a binary file with no
counts, and a tab inside a path.

No behaviour change. All three smoke suites pass, as do the end-to-end harnesses
for the trust flow, the parse cap and the layout toggle.
